+++ collector/src/main/java/com/bitmovin/analytics/adapters/PlayerAdapter.kt
@@ -2,6 +2,7 @@ package com.bitmovin.analytics.adapters
 
 import com.bitmovin.analytics.config.SourceMetadata
 import com.bitmovin.analytics.data.EventData
+import com.bitmovin.analytics.data.PlayerInfo
 import com.bitmovin.analytics.features.Feature
 import com.bitmovin.analytics.license.FeatureConfigContainer
 import com.bitmovin.analytics.stateMachines.PlayerStateMachine
@@ -11,6 +12,7 @@ interface PlayerAdapter {
     val position: Long
     val drmDownloadTime: Long?
     val currentSourceMetadata: SourceMetadata?
+    val playerInfo: PlayerInfo
 
     fun init(): Collection<Feature<FeatureConfigContainer, *>>
     fun release()