+++ collector/src/main/java/com/bitmovin/analytics/data/PlayerInfo.kt
@@ -1,0 +1,5 @@
+package com.bitmovin.analytics.data
+
+import com.bitmovin.analytics.enums.PlayerType
+
+data class PlayerInfo(val playerTech: String, val playerType: PlayerType)