+++ collector/src/test/java/com/bitmovin/analytics/retryBackend/RetryQueueTest.kt
@@ -4,6 +4,8 @@ import com.bitmovin.analytics.BitmovinAnalyticsConfig
 import com.bitmovin.analytics.data.DeviceInformation
 import com.bitmovin.analytics.data.EventData
 import com.bitmovin.analytics.data.EventDataFactory
+import com.bitmovin.analytics.data.PlayerInfo
+import com.bitmovin.analytics.enums.PlayerType
 import io.mockk.mockk
 import org.assertj.core.api.Assertions
 import org.junit.Test
@@ -24,7 +26,7 @@ class RetryQueueTest {
     }
 
     private fun setupEventData(sequenceNumber: Int): EventData {
-        var eventData = EventDataFactory(config, mockk(relaxed = true)).create("testImpressionId", null, deviceInformation)
+        var eventData = EventDataFactory(config, mockk(relaxed = true)).create("testImpressionId", null, deviceInformation, PlayerInfo("Android:Exoplayer", PlayerType.EXOPLAYER))
         eventData.sequenceNumber = sequenceNumber
         return eventData
     }