+++ collector/src/test/java/com/bitmovin/analytics/utils/DataSerializerTest.kt
@@ -9,7 +9,9 @@ import com.bitmovin.analytics.data.ErrorCode
 import com.bitmovin.analytics.data.EventData
 import com.bitmovin.analytics.data.EventDataFactory
 import com.bitmovin.analytics.data.LegacyErrorData
+import com.bitmovin.analytics.data.PlayerInfo
 import com.bitmovin.analytics.data.SecureSettingsAndroidIdUserIdProvider
+import com.bitmovin.analytics.enums.PlayerType
 import com.bitmovin.analytics.features.errordetails.ErrorData
 import io.mockk.every
 import io.mockk.mockk
@@ -29,7 +31,7 @@ class DataSerializerTest {
 
         val errorCode = ErrorCode(1000, "Error Description", ErrorData(), LegacyErrorData("Error Data Message", arrayOf("first line of details", "second line of details")))
         val bitmovinAnalyticsConfig = BitmovinAnalyticsConfig(analyticsLicenseKey, playerLicenseKey)
-        val eventData = EventDataFactory(bitmovinAnalyticsConfig, userIdProvider).create(impressionId, null, deviceInformation)
+        val eventData = EventDataFactory(bitmovinAnalyticsConfig, userIdProvider).create(impressionId, null, deviceInformation, PlayerInfo("Android:Exoplayer", PlayerType.EXOPLAYER))
 
         eventData.m3u8Url = "https://www.mydomain.com/playlist.m3u8"
         eventData.ad = 0
@@ -37,7 +39,7 @@ class DataSerializerTest {
         eventData.errorData = DataSerializer.serialize(errorCode.legacyErrorData)
         eventData.time = 1607598943236
 
-        assertThat(DataSerializer.serialize(eventData)).isEqualTo("{\"impressionId\":\"79b531da-5abb-4fb2-8dbc-9a6c60b6526f\",\"userId\":\"c54d11c8-dba2-4475-a867-764befdb5ad2\",\"key\":\"82dc5cdc-d425-4329-a043-b5fc540f9a74\",\"playerKey\":\"e32db33d-fd14-4f1c-a5e8-f95341140d04\",\"userAgent\":\"user-agent-string\",\"deviceInformation\":{\"manufacturer\":\"Google\",\"model\":\"Pixel 5\",\"isTV\":false},\"language\":\"en_US\",\"analyticsVersion\":\"${BuildConfig.COLLECTOR_CORE_VERSION}\",\"playerTech\":\"Android:Exoplayer\",\"domain\":\"package.bitmovin.com\",\"screenHeight\":640,\"screenWidth\":280,\"isLive\":false,\"isCasting\":false,\"videoDuration\":0,\"time\":1607598943236,\"videoWindowWidth\":0,\"videoWindowHeight\":0,\"droppedFrames\":0,\"played\":0,\"buffered\":0,\"paused\":0,\"ad\":0,\"seeked\":0,\"videoPlaybackWidth\":0,\"videoPlaybackHeight\":0,\"videoBitrate\":0,\"audioBitrate\":0,\"videoTimeStart\":0,\"videoTimeEnd\":0,\"videoStartupTime\":0,\"duration\":0,\"startupTime\":0,\"errorCode\":1000,\"errorData\":\"{\\\"msg\\\":\\\"Error Data Message\\\",\\\"details\\\":[\\\"first line of details\\\",\\\"second line of details\\\"]}\",\"playerStartupTime\":0,\"pageLoadType\":1,\"pageLoadTime\":0,\"m3u8Url\":\"https://www.mydomain.com/playlist.m3u8\",\"isMuted\":false,\"sequenceNumber\":0,\"platform\":\"android\",\"subtitleEnabled\":false,\"videoStartFailed\":false,\"retryCount\":0}")
+        assertThat(DataSerializer.serialize(eventData)).isEqualTo("{\"impressionId\":\"79b531da-5abb-4fb2-8dbc-9a6c60b6526f\",\"userId\":\"c54d11c8-dba2-4475-a867-764befdb5ad2\",\"key\":\"82dc5cdc-d425-4329-a043-b5fc540f9a74\",\"playerKey\":\"e32db33d-fd14-4f1c-a5e8-f95341140d04\",\"userAgent\":\"user-agent-string\",\"deviceInformation\":{\"manufacturer\":\"Google\",\"model\":\"Pixel 5\",\"isTV\":false},\"language\":\"en_US\",\"analyticsVersion\":\"${BuildConfig.COLLECTOR_CORE_VERSION}\",\"playerTech\":\"Android:Exoplayer\",\"domain\":\"package.bitmovin.com\",\"screenHeight\":640,\"screenWidth\":280,\"isLive\":false,\"isCasting\":false,\"videoDuration\":0,\"time\":1607598943236,\"videoWindowWidth\":0,\"videoWindowHeight\":0,\"droppedFrames\":0,\"played\":0,\"buffered\":0,\"paused\":0,\"ad\":0,\"seeked\":0,\"videoPlaybackWidth\":0,\"videoPlaybackHeight\":0,\"videoBitrate\":0,\"audioBitrate\":0,\"videoTimeStart\":0,\"videoTimeEnd\":0,\"videoStartupTime\":0,\"duration\":0,\"startupTime\":0,\"errorCode\":1000,\"errorData\":\"{\\\"msg\\\":\\\"Error Data Message\\\",\\\"details\\\":[\\\"first line of details\\\",\\\"second line of details\\\"]}\",\"playerStartupTime\":0,\"pageLoadType\":1,\"pageLoadTime\":0,\"m3u8Url\":\"https://www.mydomain.com/playlist.m3u8\",\"isMuted\":false,\"sequenceNumber\":0,\"platform\":\"android\",\"subtitleEnabled\":false,\"videoStartFailed\":false,\"retryCount\":0,\"player\":\"exoplayer\"}")
     }
 
     @Test
@@ -60,14 +62,14 @@ class DataSerializerTest {
 
         val errorCode = ErrorCode(1000, "Error Description", ErrorData(), null)
         val bitmovinAnalyticsConfig = BitmovinAnalyticsConfig(analyticsLicenseKey, playerLicenseKey)
-        val eventData = EventDataFactory(bitmovinAnalyticsConfig, userIdProvider).create(impressionId, null, deviceInformation)
+        val eventData = EventDataFactory(bitmovinAnalyticsConfig, userIdProvider).create(impressionId, null, deviceInformation, PlayerInfo("Android:Exoplayer", PlayerType.EXOPLAYER))
 
         eventData.m3u8Url = "https://www.mydomain.com/playlist.m3u8"
         eventData.ad = 0
         eventData.errorCode = errorCode.errorCode
         eventData.errorData = DataSerializer.serialize(null)
         eventData.time = 1607598943236
-        assertThat(DataSerializer.serialize(eventData)).isEqualTo("{\"impressionId\":\"79b531da-5abb-4fb2-8dbc-9a6c60b6526f\",\"userId\":\"c54d11c8-dba2-4475-a867-764befdb5ad2\",\"key\":\"82dc5cdc-d425-4329-a043-b5fc540f9a74\",\"playerKey\":\"e32db33d-fd14-4f1c-a5e8-f95341140d04\",\"userAgent\":\"user-agent-string\",\"deviceInformation\":{\"manufacturer\":\"Google\",\"model\":\"Pixel 5\",\"isTV\":false},\"language\":\"en_US\",\"analyticsVersion\":\"${BuildConfig.COLLECTOR_CORE_VERSION}\",\"playerTech\":\"Android:Exoplayer\",\"domain\":\"package.bitmovin.com\",\"screenHeight\":640,\"screenWidth\":280,\"isLive\":false,\"isCasting\":false,\"videoDuration\":0,\"time\":1607598943236,\"videoWindowWidth\":0,\"videoWindowHeight\":0,\"droppedFrames\":0,\"played\":0,\"buffered\":0,\"paused\":0,\"ad\":0,\"seeked\":0,\"videoPlaybackWidth\":0,\"videoPlaybackHeight\":0,\"videoBitrate\":0,\"audioBitrate\":0,\"videoTimeStart\":0,\"videoTimeEnd\":0,\"videoStartupTime\":0,\"duration\":0,\"startupTime\":0,\"errorCode\":1000,\"playerStartupTime\":0,\"pageLoadType\":1,\"pageLoadTime\":0,\"m3u8Url\":\"https://www.mydomain.com/playlist.m3u8\",\"isMuted\":false,\"sequenceNumber\":0,\"platform\":\"android\",\"subtitleEnabled\":false,\"videoStartFailed\":false,\"retryCount\":0}")
+        assertThat(DataSerializer.serialize(eventData)).isEqualTo("{\"impressionId\":\"79b531da-5abb-4fb2-8dbc-9a6c60b6526f\",\"userId\":\"c54d11c8-dba2-4475-a867-764befdb5ad2\",\"key\":\"82dc5cdc-d425-4329-a043-b5fc540f9a74\",\"playerKey\":\"e32db33d-fd14-4f1c-a5e8-f95341140d04\",\"userAgent\":\"user-agent-string\",\"deviceInformation\":{\"manufacturer\":\"Google\",\"model\":\"Pixel 5\",\"isTV\":false},\"language\":\"en_US\",\"analyticsVersion\":\"${BuildConfig.COLLECTOR_CORE_VERSION}\",\"playerTech\":\"Android:Exoplayer\",\"domain\":\"package.bitmovin.com\",\"screenHeight\":640,\"screenWidth\":280,\"isLive\":false,\"isCasting\":false,\"videoDuration\":0,\"time\":1607598943236,\"videoWindowWidth\":0,\"videoWindowHeight\":0,\"droppedFrames\":0,\"played\":0,\"buffered\":0,\"paused\":0,\"ad\":0,\"seeked\":0,\"videoPlaybackWidth\":0,\"videoPlaybackHeight\":0,\"videoBitrate\":0,\"audioBitrate\":0,\"videoTimeStart\":0,\"videoTimeEnd\":0,\"videoStartupTime\":0,\"duration\":0,\"startupTime\":0,\"errorCode\":1000,\"playerStartupTime\":0,\"pageLoadType\":1,\"pageLoadTime\":0,\"m3u8Url\":\"https://www.mydomain.com/playlist.m3u8\",\"isMuted\":false,\"sequenceNumber\":0,\"platform\":\"android\",\"subtitleEnabled\":false,\"videoStartFailed\":false,\"retryCount\":0,\"player\":\"exoplayer\"}")
     }
 
     @Test