+++ CHANGELOG.md
@@ -19,8 +19,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - [internal] updated gradle plugin and dependencies
 - [internal] updated spotless config
 - [internal] move to main as default branch
+- [internal] refactored PlayerTech to be constant and part of adapters
 - Upgraded gson dependency to v2.8.9
 - Upgraded example apps to latest bitmovin player (v3.29.0) and latest exoplayer (v2.18.2)
+- Deprecated obsolete PlayerType config in AnalyticsConfig
 
 ### Fixed
 - Allow exoplayer >= v2.17.0 to be used with the collector